Network Improvements (#44)
* Add documentation

* Make non-throwing

* Remove force cast

* Use more explicit name

* Move into separate file

* Reorganize

* Declare Codable conformance
