+++ Sources/BlueTriangle/Constants.swift
@@ -12,13 +12,16 @@ enum Constants {
     static let device = "Mobile"
     static let os = "iOS"
     static let globalUserIDKey = "com.bluetriangle.kGlobalUserIDUserDefault"
-    static let timerEndpoint: URL = "https://d.btttag.com/analytics.rcv"
-    static let capturedRequestEndpoint: URL = "https://d.btttag.com/wcdv02.rcv"
-    static let errorEndpoint: URL = "https://d.btttag.com/err.rcv"
 
+    // Settings
     static let minimumSampleInterval: TimeInterval = 1 / 60
     static let sessionTimeoutInMinutes = 30
     static let userSessionTimeoutInDays = 365
+
+    // Endpoints
+    static let capturedRequestEndpoint: URL = "https://d.btttag.com/wcdv02.rcv"
+    static let errorEndpoint: URL = "https://d.btttag.com/err.rcv"
+    static let timerEndpoint: URL = "https://d.btttag.com/analytics.rcv"
 
     // Crash Tracking
     static let crashID = "iOS Crash"
